move serve_down to core and refactor
