+++ sky/__init__.py
@@ -18,6 +18,7 @@ from sky.core import down
 from sky.core import download_logs
 from sky.core import job_status
 from sky.core import queue
+from sky.core import serve_down
 from sky.core import spot_cancel
 from sky.core import spot_queue
 from sky.core import spot_status
@@ -33,7 +34,6 @@ from sky.data import StorageMode
 from sky.data import StoreType
 from sky.execution import exec  # pylint: disable=redefined-builtin
 from sky.execution import launch
-from sky.execution import serve_down
 from sky.execution import serve_up
 from sky.execution import spot_launch
 from sky.optimizer import Optimizer
