+++ sky/execution.py
@@ -1193,128 +1193,3 @@ def serve_up(
               f'{handle.head_ip}:{load_balancer_port}{style.RESET_ALL}')
         print(f'{fore.GREEN}Starting replicas now...{style.RESET_ALL}')
         print('Please use the above command to find the latest status.')
-
-
-def serve_down(
-    service_name: str,
-    purge: bool,
-):
-    """Teardown a service.
-
-    Please refer to the sky.cli.serve_down for the document.
-
-    Args:
-        service_name: Name of the service.
-        purge: If true, ignore errors when cleaning up the controller.
-    """
-    service_record = global_user_state.get_service_from_name(service_name)
-
-    if service_record is None:
-        with ux_utils.print_exception_no_traceback():
-            raise ValueError(f'Service {service_name!r} not found.')
-
-    service_handle: serve.ServiceHandle = service_record['handle']
-    controller_name = service_record['controller_name']
-    global_user_state.set_service_status(service_name,
-                                         status_lib.ServiceStatus.SHUTTING_DOWN)
-    handle = global_user_state.get_handle_from_cluster_name(controller_name)
-
-    if handle is not None:
-        backend = backend_utils.get_backend_from_handle(handle)
-        assert isinstance(backend, backends.CloudVmRayBackend)
-        try:
-            if service_handle.controller_port is None:
-                with ux_utils.print_exception_no_traceback():
-                    raise RuntimeError(
-                        f'Controller job of service {service_name!r} not found.'
-                    )
-
-            code = serve.ServeCodeGen.terminate_service(
-                service_handle.controller_port)
-            returncode, terminate_service_payload, stderr = backend.run_on_head(
-                handle,
-                code,
-                require_outputs=True,
-                stream_logs=False,
-                separate_stderr=True)
-            subprocess_utils.handle_returncode(
-                returncode,
-                code, ('Failed when submit terminate request to controller '
-                       f'of service {service_name!r}'),
-                stderr,
-                stream_logs=False)
-
-            resp = serve.load_terminate_service_result(
-                terminate_service_payload)
-            if resp.status_code != 200:
-                raise RuntimeError('Failed to terminate replica of service '
-                                   f'{service_name!r} due to request '
-                                   f'failure: {resp.text}')
-            msg = resp.json()['message']
-            if msg:
-                raise RuntimeError(
-                    'Unexpected message when tearing down replica of service '
-                    f'{service_name!r}: {msg}. Please login to the controller '
-                    'and make sure the service is properly cleaned.')
-
-        # We want to make sure no matter what error happens, we can still
-        # clean up the record if purge is True.
-        except Exception as e:  # pylint: disable=broad-except
-            if purge:
-                logger.warning('Ignoring error when cleaning replicas of '
-                               f'{service_name!r}: {e}')
-            else:
-                raise RuntimeError(e) from e
-    else:
-        if not purge:
-            with ux_utils.print_exception_no_traceback():
-                raise RuntimeError(
-                    f'Cannot find controller of service {service_name!r}.')
-
-    try:
-        if handle is not None:
-            assert isinstance(handle, backends.CloudVmRayResourceHandle)
-            backend = backends.CloudVmRayBackend()
-
-            # Cancel the controller and load balancer jobs.
-            # For the case when controller / load_balancer job failed to submit.
-            jobs = []
-            if service_handle.job_id is not None:
-                jobs.append(service_handle.job_id)
-            backend.cancel_jobs(handle, jobs=jobs, silent=True)
-
-            # Cleanup all files on controller related to this service.
-            # We have a 10-min grace period for the controller to autostop,
-            # so it should be fine if this is the last service on the
-            # controller and its job is the only one running.
-            code = serve.ServeCodeGen.cleanup_service_files(service_name)
-            returncode, _, stderr = backend.run_on_head(handle,
-                                                        code,
-                                                        require_outputs=True,
-                                                        stream_logs=False,
-                                                        separate_stderr=True)
-            subprocess_utils.handle_returncode(
-                returncode,
-                code, ('Failed when cleaning up service files on controller '
-                       f'of service {service_name!r}'),
-                stderr,
-                stream_logs=False)
-
-    # same as above.
-    except Exception as e:  # pylint: disable=broad-except
-        if purge:
-            logger.warning(
-                'Ignoring error when stopping controller and '
-                f'load balancer jobs of service {service_name!r}: {e}')
-        else:
-            raise RuntimeError(e) from e
-
-    # TODO(tian): Maybe add a post_cleanup function?
-    controller_yaml_path = serve.generate_controller_yaml_file_name(
-        service_name)
-    if os.path.exists(controller_yaml_path):
-        os.remove(controller_yaml_path)
-    handle = global_user_state.get_handle_from_service_name(service_name)
-    assert handle is not None
-    handle.cleanup_ephemeral_storage()
-    global_user_state.remove_service(service_name)
